首页推荐游戏完善

+++ main/src/main/java/com/gamersky/kagurasansan/data/bean/ChannelListData.java
@@ -37,7 +37,7 @@ public class ChannelListData extends DataBaseBean {
 
         public String type;
         public String contentType;
-        public int contentId;
+        public long contentId;
         public String title;
         public String description;
         public List<String> thumbnailURLs;
